% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admin1.map}
\alias{admin1.map}
\title{An sf containing geometry data for sub-country regions}
\usage{
data(admin1.map)
}
\description{
An sf containing geometry data for sub-country regions
}
\references{
Data obtained using rnatrualearth
}
