% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admin1.regions}
\alias{admin1.regions}
\title{Supported world sub-country regions with lookup data in multiple languages}
\usage{
data(admin1.regions)
}
\description{
This dataset contains feature names for world sub-country regions in many
languages and formats, and can be used to join your data to the identifiers
accepted by the choropleth_admin1() function. For feature names in even more
languages, please see rnaturalearth::ne_states().
}
\references{
Data obtained using rnatrualearth
}
