% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Admin1Choropleth}
\alias{Admin1Choropleth}
\title{An R6 object for creating Administration Level 1 choropleths.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7f9d9333e208> 
 - attr(*, "name")= chr "Admin1Choropleth_generator"
}}
\usage{
Admin1Choropleth
}
\description{
An R6 object for creating Administration Level 1 choropleths.
}
\keyword{datasets}

