\docType{data}
\name{all.us.states.map.df}
\alias{all.us.states.map.df}
\title{Map of the 50 US states plus the district of columbia.}
\description{
A data.frame which contains a map of all 50 US States plus
the District of Columbia.  Taken from the US Census 2010
Cartographic Boundary shapefiles
(https://www.census.gov/geo/maps-data/data/tiger-cart-boundary.html).
The resolutions is 20m (20m = 1:20,000,000). The shapefile
was modified using the QGIS program in order to 1) remove
Puerto Rico and 2) remove islands off of Alaska that
crossed the antimeridian.
}
\keyword{data}

