% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_percent_change}
\alias{calculate_percent_change}
\title{Calculate the percentage change between two choroplethr dataframes.}
\usage{
calculate_percent_change(df1, df2)
}
\arguments{
\item{df1}{A dataframe with columns named "region" and "value"}

\item{df2}{A dataframe with columns named "region" and "value"}
}
\description{
Merges df1 and df2 on column named "region", and computes percentage change from
df1$value to df2$value. Result is in the new "value" column, and rounded to two
digits.
}
