% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{Choropleth}
\alias{Choropleth}
\title{The base Choropleth object.}
\description{
The base Choropleth object.

The base Choropleth object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Choropleth-new}{\code{Choropleth$new()}}
\item \href{#method-Choropleth-set_zoom}{\code{Choropleth$set_zoom()}}
\item \href{#method-Choropleth-get_ggscale}{\code{Choropleth$get_ggscale()}}
\item \href{#method-Choropleth-get_projection}{\code{Choropleth$get_projection()}}
\item \href{#method-Choropleth-render}{\code{Choropleth$render()}}
\item \href{#method-Choropleth-clone}{\code{Choropleth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth-new"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$new(
  ref.regions,
  ref.regions.name,
  map.df,
  geoid.all,
  user.df,
  geoid.name,
  geoid.type,
  value.name,
  num_colors,
  label_col
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth-set_zoom"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth-set_zoom}{}}}
\subsection{Method \code{set_zoom()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$set_zoom(zoom)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth-get_ggscale"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth-get_ggscale}{}}}
\subsection{Method \code{get_ggscale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$get_ggscale(
  choropleth.df = self$choropleth.df,
  respect_zoom = TRUE,
  custom.colors,
  color.min,
  color.max,
  na.color,
  nbreaks
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth-get_projection"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth-get_projection}{}}}
\subsection{Method \code{get_projection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$get_projection(
  choropleth.df = self$choropleth.df,
  respect_zoom = TRUE,
  projection_name,
  ignore_latlon,
  limits_lat,
  limits_lon,
  reproject,
  whitespace
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth-render"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth-render}{}}}
\subsection{Method \code{render()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$render(
  choropleth.df = self$choropleth.df,
  ggscale,
  projection,
  respect_zoom = TRUE,
  occlude_latlon_limits,
  border_color,
  border_thickness,
  background_color,
  gridlines,
  latlon_ticks,
  label,
  label_text_size,
  label_text_color,
  label_box_color,
  ggrepel_options,
  legend,
  legend_position,
  title,
  addl_gglayer
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
