% Generated by roxygen2 (4.0.2): do not edit by hand
\name{choroplethr_wdi}
\alias{choroplethr_wdi}
\title{Create a country-level choropleth using data from the World Bank Development Indicators (WDI).}
\usage{
choroplethr_wdi(code = "SP.POP.TOTL", year = 2012, title = NULL,
  countries = NULL, num_buckets = 7)
}
\arguments{
\item{code}{The WDI code to use.}

\item{year}{The year of data to use.}

\item{title}{A title for the map.  If not specified, automatically generated to include WDI code and year.}

\item{countries}{If null, render all countries. Otherwise must be a vector of country names.  See ?country.names for a list of valid country names.}

\item{num_buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.  For
example, 2 will show values above or below the median, and 9 will show the maximum
resolution.  Defaults to 9.}
}
\value{
A choropleth.
}
\description{
Create a country-level choropleth using data from the World Bank Development Indicators (WDI).
}
\examples{
\dontrun{
# See http://data.worldbank.org/indicator/SP.POP.TOTL
choroplethr_wdi(code="SP.POP.TOTL", year=2012, title="2012 Per Capita Income", num_buckets=1)

# See http://data.worldbank.org/indicator/SP.DYN.LE00.IN
choroplethr_wdi(code="SP.DYN.LE00.IN", year=2012, title="2012 Life Expectancy Estimates")

# See http://data.worldbank.org/indicator/NY.GDP.PCAP.CD
choroplethr_wdi(code="NY.GDP.PCAP.CD", year=2012, title="2012 Per Capita Income")
}
}
\references{
Uses the WDI function from the WDI package by Vincent Arel-Bundock.
}

