% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{congress116.regions}
\alias{congress116.regions}
\title{A data.frame containing geographic metadata about the Congressional Districts of the 116th US Congress}
\usage{
data(congress116.regions)
}
\description{
Column region is how the Census Bureau refers to the geography. Note that this region is a 4-character string,
and so has a leading 0 if necessary. The first two characters are the state FIPS code, and the second two characters
are the district ID. States that only have 1 district (i.e. a representative "at large") have district 00. 
All other states start at 01.
}
\keyword{data}
