% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{continental_us_states}
\alias{continental_us_states}
\title{A vector of the names of US Continental US States.}
\usage{
data(continental_us_states)
}
\description{
A vector of the names of US Continental US States.
}
\author{
Ari Lamstein
}
\keyword{data}
