% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country.map}
\alias{country.map}
\title{An sf containing geometry data for countries of the world}
\usage{
data(country.map)
}
\description{
An sf containing geometry data for countries of the world
}
\references{
Data obtained using the ne_countries function from rnaturalearth; \cr
  \url{https://github.com/ropensci/rnaturalearth}, \cr
  \url{https://www.naturalearthdata.com/}
}
