% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{country.names}
\alias{country.names}
\title{Names of all counties on the country.map data.frame. A data.frame that includes both English names and
their iso2c equivalents.}
\usage{
data(country.names)
}
\description{
Names of all counties on the country.map data.frame. A data.frame that includes both English names and
their iso2c equivalents.
}

