% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country.regions}
\alias{country.regions}
\title{Supported countries along with lookup data in multiple languages}
\usage{
data(country.regions)
}
\description{
Supported countries along with lookup data in multiple languages
}
