% Generated by roxygen2 (4.0.1): do not edit by hand
\name{country_wdi}
\alias{country_wdi}
\title{Create a country choropleth using data from the World Bank Development Indicators (WDI).}
\usage{
country_wdi(code = "NY.GNS.ICTR.GN.ZS", year = 2012, title = NULL,
  countries = NULL, num_buckets = 7)
}
\arguments{
\item{code}{The WDI code to use.}

\item{year}{The year of data to use.}

\item{title}{A title for the map.  If not specified, automatically generated to include WDI code and year.}

\item{countries}{If null, render all countries. Otherwise must be a vector of country names.  See ?country.names for a list of valid country names.
Use the column "region" not "iso2c".}

\item{num_buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.  For
example, 2 will show values above or below the median, and 9 will show the maximum
resolution.  Defaults to 7.}
}
\value{
A choropleth.
}
\description{
Create a country choropleth using data from the World Bank Development Indicators (WDI).
}

