% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{CountryChoropleth}
\alias{CountryChoropleth}
\title{An R6 object for creating country-level choropleths.}
\format{\preformatted{List of 8
 $ new       :function (...)  
 $ classname : chr "CountryChoropleth"
 $ public    :List of 20
  ..$ user.df        : NULL
  ..$ map.df         : NULL
  ..$ choropleth.df  : NULL
  ..$ title          : chr ""
  ..$ legend_name    : chr ""
  ..$ warn           : logi TRUE
  ..$ ggplot_scale   : NULL
  ..$ render         :function ()  
  ..$ clip           :function ()  
  ..$ discretize     :function ()  
  ..$ bind           :function ()  
  ..$ prepare_map    :function ()  
  ..$ get_scale      :function ()  
  ..$ theme_clean    :function ()  
  ..$ theme_inset    :function ()  
  ..$ format_levels  :function (x, nsep = " to ")  
  ..$ set_zoom       :function (zoom)  
  ..$ get_zoom       :function ()  
  ..$ set_num_buckets:function (num_buckets)  
  ..$ initialize     :function (user.df)  
 $ private   :List of 2
  ..$ zoom       : NULL
  ..$ num_buckets: num 7
 $ active    : NULL
 $ inherit   :List of 8
  ..$ new       :function (...)  
  ..$ classname : chr "Choropleth"
  ..$ public    :List of 20
  .. ..$ user.df        : NULL
  .. ..$ map.df         : NULL
  .. ..$ choropleth.df  : NULL
  .. ..$ title          : chr ""
  .. ..$ legend_name    : chr ""
  .. ..$ warn           : logi TRUE
  .. ..$ ggplot_scale   : NULL
  .. ..$ initialize     :function (map.df, user.df)  
  .. ..$ render         :function ()  
  .. ..$ clip           :function ()  
  .. ..$ discretize     :function ()  
  .. ..$ bind           :function ()  
  .. ..$ prepare_map    :function ()  
  .. ..$ get_scale      :function ()  
  .. ..$ theme_clean    :function ()  
  .. ..$ theme_inset    :function ()  
  .. ..$ format_levels  :function (x, nsep = " to ")  
  .. ..$ set_zoom       :function (zoom)  
  .. ..$ get_zoom       :function ()  
  .. ..$ set_num_buckets:function (num_buckets)  
  ..$ private   :List of 2
  .. ..$ zoom       : NULL
  .. ..$ num_buckets: num 7
  ..$ active    : NULL
  ..$ inherit   : NULL
  ..$ parent_env:<environment: 0x7fd70843b5c8> 
  ..$ lock      : logi TRUE
  ..- attr(*, "class")= chr "R6ClassGenerator"
 $ parent_env:<environment: 0x7fd70843b5c8> 
 $ lock      : logi TRUE
 - attr(*, "class")= chr "R6ClassGenerator"
}}
\usage{
CountryChoropleth
}
\description{
An R6 object for creating country-level choropleths.
}
\keyword{datasets}

