% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county.map.2015}
\alias{county.map.2015}
\title{An sf containing geometry data for US counties in 2015}
\usage{
data(county.map.2015)
}
\description{
An sf containing geometry data for US counties in 2015
}
\references{
obtained using tigris::counties()
}
