% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county.map.2024}
\alias{county.map.2024}
\title{An sf containing geometry data for US counties in 2024}
\usage{
data(county.map.2024)
}
\description{
An sf containing geometry data for US counties in 2024
}
\references{
obtained using tigris::counties()
}
