% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county.regions.2024}
\alias{county.regions.2024}
\title{Supported regions for US counties in 2024}
\usage{
data(county.regions.2024)
}
\description{
Supported regions for US counties in 2024
}
