% Generated by roxygen2 (4.0.2): do not edit by hand
\name{county_zoom_choropleth}
\alias{county_zoom_choropleth}
\title{Create a choropleth of USA Counties, with sensible defaults, that zooms on counties.}
\usage{
county_zoom_choropleth(df, title = "", legend = "", buckets = 7,
  zoom = NULL)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".  Elements in
the "region" column must exactly match how regions are named in the "region" column in county.map.}

\item{title}{An optional title for the map.}

\item{legend}{An optional name for the legend.}

\item{buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.}

\item{zoom}{An optional vector of states to zoom in on. Elements of this vector must exactly
match the names of states as they appear in the "region" column of ?state.regions.}
}
\description{
The map used is county.map in the choroplethrMaps package.  See country.regions
in the choroplethrMaps package for an object which can help you coerce your regions
into the required format.
}
\examples{
\dontrun{

# show the population of the 5 counties (boroughs) that make up New York City.

nyc_fips=c(36005, 36047, 36081, 36061, 36085)
 county_zoom_choropleth(df_pop_county,
                        zoom=nyc_fips,
                        buckets=1,
                        title="Population of Counties in New York City",
                        scale="Population")
}
}

