% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_congress116_demographics}
\alias{df_congress116_demographics}
\title{A data.frame containing demographic statistics about the 116th Congressional Districts}
\usage{
data(df_congress116_demographics)
}
\description{
A data.frame containing demographic statistics about the 116th Congressional Districts
}
\references{
Data comes from the 2018 5-year American Community Survey (ACS). Data generated by
?get_congressional_district_demographics
}
\keyword{data}
