% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_congress116_party}
\alias{df_congress116_party}
\title{A data.frame containing party affiliation data about the Congressional Districts of 116th US Congress}
\usage{
data(df_congress116_party)
}
\description{
Contains the party affiliation of each member elected to the House of Representatives of the 116th Congress, along with metadata.
Note that party affiliation is of who the citizens voted for, and not who is currently 
(July 30, 2020) serving. Currently three members have resigned since being elected, one 
switched party and one died. For details of how this data was compiled, please see function 
get_congressional_116_party_data in file get_congress_116_party_data. That file ships with this package, 
but is not exported, since it relies on scraping data from Wikipedia, and that web page is subject 
to change.
}
\keyword{data}
