% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_congress_116_regions}
\alias{df_congress_116_regions}
\title{A data.frame containing geographic metadata about the 116th Congressional Congress}
\usage{
data(df_congress_116_regions)
}
\description{
Column region is how the Census Bureau refers to the geography. Note that this region is a 4-character string,
and so has a leading 0 if necessary. The first two characters are the state FIPS code, and the second two characters
are the district ID. States that only have 1 district (i.e. a representative "at large") have district 00. 
All other states seem to start at 01.
}
\details{
In addition to the region column, this data.frame has the party the region voted for, as well 
as other metadata about the District.
}
\author{
Ari Lamstein
}
\keyword{data}
