% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_country_demographics}
\alias{df_country_demographics}
\title{A data.frame containing population estimates for Countries in 2012.}
\usage{
data(df_country_demographics)
}
\description{
A data.frame containing population estimates for Countries in 2012.
}
\references{
Data obtained using the ne_countries function from rnaturalearth \cr
  \url{https://github.com/ropensci/rnaturalearth},\cr
  \url{https://www.naturalearthdata.com/}
}
\keyword{data}
