% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_county_demographics}
\alias{df_county_demographics}
\title{A data.frame containing demographic statistics for each county in the United
States.}
\usage{
data(df_county_demographics)
}
\description{
A data.frame containing demographic statistics for each county in the United
States.
}
\references{
Data comes from the 2013 5-year American Community Survey (ACS).\cr
  Data generated by get_county_demographics().
}
\keyword{data}
