% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_ny_tract_demographics}
\alias{df_ny_tract_demographics}
\title{A data.frame containing demographic statistics for each Census Tract in New York State.}
\usage{
data(df_ny_tract_demographics)
}
\description{
A data.frame containing demographic statistics for each Census Tract in New York State.
}
\references{
Data comes from the 2013 5-year American Community Survey (ACS). Data generated by
?get_tract_demographics.
}
\keyword{data}
