% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pop_ny_tract}
\alias{df_pop_ny_tract}
\title{A data.frame containing population estimates for all Census Tracts in New York 
State in 2012.}
\usage{
data(df_pop_ny_tract)
}
\description{
A data.frame containing population estimates for all Census Tracts in New York 
State in 2012.
}
\references{
Taken from the US American Community Survey (ACS) 5 year estimates.
}
\keyword{data}
