% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zip_data.R
\docType{data}
\name{df_pop_zip}
\alias{df_pop_zip}
\title{A data.frame containing population estimates for US Zip Code Tabulated Areas (ZCTAs) in 2012.}
\usage{
data(df_pop_zip)
}
\description{
ZCTAs are intended to be roughly analogous to postal ZIP codes.
}
\examples{
data(df_pop_zip)
zip_map(df_pop_zip, title="2012 Population Estimates")
}
\references{
Taken from the US American Community Survey (ACS) 5 year estimates.
ZCTAs, and their realationship to ZIP codes, are explained here \url{https://www.census.gov/geo/reference/zctas.html}.
}
\keyword{data}

