\docType{data}
\name{df_president_2012}
\alias{df_president_2012}
\title{A data.frame containing election results from the 2012 US Presidential election.}
\description{
A data.frame containing election results from the 2012 US
Presidential election.
}
\examples{
data(df_president_2012)
choroplethr(df_president_2012, "state", title="Results from the 2012 US Presidential Election")
}
\author{
US Federal Election Commission and Ari Lamstein
}
\references{
Taken from
\url{http://www.fec.gov/pubrec/fe2012/federalelections2012.shtml}
on 3/15/2014.
}
\keyword{data}

