% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_president_margin}
\alias{df_president_margin}
\title{A data.frame containing election result percentage margins from the 2012 US Presidential election.}
\usage{
data(df_president_margin)
}
\description{
A data.frame containing election result percentage margins from the 2012 US Presidential election.
}
\details{
Negative values are percent for Romney, positive values are percent for Obama.
}
\author{
Katrina Kutchko
}
\references{
Taken from the FEC website on 12/6/2016.
}
\keyword{data}

