% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_ri_zip_demographics}
\alias{df_ri_zip_demographics}
\title{A data.frame containing demographic statistics for each zip code in Rhode
Island.}
\usage{
data(df_ri_zip_demographics)
}
\description{
A data.frame containing demographic statistics for each zip code in Rhode
Island.
}
\references{
Data comes from the 2013 5-year American Community Survey (ACS).\cr
}
\keyword{data}
