% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_state_age_2015}
\alias{df_state_age_2015}
\title{A data.frame containing median age estimates for US states in 2015}
\usage{
data(df_state_age_2015)
}
\description{
A data.frame containing median age estimates for US states in 2015
}
\references{
Taken from the US American Community Survey (ACS) 5 year estimates.
}
\keyword{data}
