% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_state_demographics}
\alias{df_state_demographics}
\title{A data.frame containing demographic statistics for each state plus the District of Columbia.}
\usage{
data(df_state_demographics)
}
\description{
A data.frame containing demographic statistics for each state plus the District of Columbia.
}
\references{
Data comes from the 2013 5-year American Community Survey (ACS). Data generated by
?get_state_demographics.
}
\keyword{data}

