% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{double_map}
\alias{double_map}
\title{Place two maps side by side}
\usage{
double_map(map1, map2, title = "")
}
\arguments{
\item{map1}{The first map}

\item{map2}{The second map}

\item{title}{An optional title}
}
\description{
With an optional title. Especially useful for contrasting choropleth maps both with
and without a reference map underneath.
}
