% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_acs_column_names}
\alias{get_acs_column_names}
\title{Returns the column names of an ACS table.}
\usage{
get_acs_column_names(tableId, lod, endyear = 2012, span = 5)
}
\arguments{
\item{tableId}{The id of an ACS table.}

\item{lod}{The level of geographic detail for the data.frame.  Must be one of "state", "county" or "zip".}

\item{endyear}{The end year of the survey.  Defaults to 2012.}

\item{span}{The span of the survey.  Defaults to 5.}
}
\value{
A vector of column names
}
\description{
Returns the column names of an ACS table.
}
\seealso{
http://factfinder2.census.gov/faces/help/jsf/pages/metadata.xhtml?lang=en&type=survey&id=survey.en.ACS_ACS, which lists all ACS Surveys.
}

