% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_congressional_demographics.R
\name{get_congressional_district_demographics}
\alias{get_congressional_district_demographics}
\title{Get a handful of demographic variables on US Congressional Districts from the US Census Bureau as a data.frame.}
\usage{
get_congressional_district_demographics(year = 2018, survey = "acs5")
}
\arguments{
\item{year}{The year the survey was published}

\item{survey}{The survey. Either "acs5" or "acs1"}
}
\description{
The data comes from the American Community Survey (ACS). The variables are: total population, percent White 
not Hispanic, Percent Black or African American not Hispanic, percent Asian not Hispanic,
percent Hispanic all races, per-capita income, median rent and median age.
}
