% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usa_demographics.R
\name{get_county_demographics}
\alias{get_county_demographics}
\title{Get a handful of demographic variables on US Counties from the US Census Bureau as a data.frame.}
\usage{
get_county_demographics(endyear = 2013, span = 5)
}
\arguments{
\item{endyear}{The end year for the survey}

\item{span}{The span of the survey}
}
\description{
The data comes from the American Community Survey (ACS). The variables are total population
and median household income.
}
