% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usa_demographics.R
\name{get_tract_demographics}
\alias{get_tract_demographics}
\title{Get a handful of demographic variables on Census Tracts in a State from the US Census Bureau as a data.frame.}
\usage{
get_tract_demographics(
  state_name,
  county_fips = NULL,
  endyear = 2013,
  span = 5
)
}
\arguments{
\item{state_name}{The name of the state. See ?state.regions for proper spelling and capitalization.}

\item{county_fips}{An optional vector of county fips codes within the state. Useful to set because getting data on all tracts can be slow.}

\item{endyear}{The end year for the survey}

\item{span}{The span of the survey}
}
\description{
The data comes from the American Community Survey (ACS). The variables are total population
and median household income.
}
\examples{
\donttest{
# 36061 is the FIPS code for Manhatttan (technically "New York County"), NY.
df = get_tract_demographics("new york", 36061)
df$value = df$median_hh_income
tract_choropleth(df, "new york", county_zoom = 36061) 
}
}
