% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract.R
\name{get_tract_map}
\alias{get_tract_map}
\title{Download a map of all census tracts in a given state}
\usage{
get_tract_map(state_name, map_year, drop_geometry = TRUE)
}
\arguments{
\item{state_name}{The name of the state, given by proper name, abbreviation,
for FIPS code.}

\item{map_year}{The year of the map you want to download.}

\item{drop_geometry}{Drop geometry data?}
}
\description{
The map returned is exactly the same map which tract_choropleth uses. It is
downloaded using the "tracts" function in the tigris package, and then it is
modified for use with choroplethr.
}
