% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{get_zip_map}
\alias{get_zip_map}
\title{Download a map of requested zip codes}
\usage{
get_zip_map(zipcodes, map_year, drop_geometry = TRUE)
}
\arguments{
\item{zipcodes}{The zipcodes for which you want to download map data.}

\item{map_year}{The year of the map you want to download. Only years after 
2013 are available.}

\item{drop_geometry}{Drop geometry data?}
}
\description{
The map returned is exactly the same map which zip_choropleth() uses. It is
downloaded using the "zctas" function in the tigris package, and then it is
modified for use with choroplethr.
}
