% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zip_choropleth.R
\name{install_choroplethrZip}
\alias{install_choroplethrZip}
\title{Installs the choroplethrZip package from github}
\usage{
install_choroplethrZip()
}
\description{
The choroplethrZip package contains shapefiles and metadata of US Zip Code Tabulated Areas
from the US Census Bureau. Due to the package's size (64MB) it is hosted on github, not CRAN.
The address of the respository is https://github.com/arilamstein/choroplethrZip.
}

