% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{map.world}
\alias{map.world}
\title{Map of the world.}
\description{
This data.frame corresponds to version 2.0.0 of the "Admin 0 - Countries" map from naturalearthdata.com
The data.frame was modified by removing columns with non-ASCII characters. Also,
I added a column called "region" which is the the all lowercase version of the
column "sovereignt".
}
\references{
Taken from http://www.naturalearthdata.com/downloads/110m-cultural-vectors/
}

