% Generated by roxygen2 (4.0.1): do not edit by hand
\name{percent_change}
\alias{percent_change}
\title{Given two choropleth dfs, return a df that represents the percent change}
\usage{
percent_change(a, b)
}
\arguments{
\item{a}{A data.frame representing a choropleth, with one column named region and one column named value}

\item{b}{A data.frame representing a choropleth, with one column named region and one column named value}
}
\description{
Given two choropleth dfs, return a df that represents the percent change
}

