% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simple_state_choropleth}
\alias{simple_state_choropleth}
\title{Create a simple choropleth of USA States}
\usage{
simple_state_choropleth(df, title = "", legend_name = "", num_buckets = 7)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".}

\item{title}{An optional title for the map.}

\item{legend_name}{An optional name for the legend.}

\item{num_buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.}
}
\description{
AK and HI not rendered as insets
}
\examples{
data(df_pop_state)
simple_state_choropleth(df_pop_state, title="US 2012 Population Estimates", legend_name="Population")
}

