% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.map.bigdc}
\alias{state.map.bigdc}
\title{An sf containing geometry data for US states with DC enlarged}
\usage{
data(state.map.bigdc)
}
\description{
An sf containing geometry data for US states with DC enlarged
}
