% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.map.hex}
\alias{state.map.hex}
\title{An sf containing a hexagonal tile map for US states}
\usage{
data(state.map.hex)
}
\description{
An sf containing a hexagonal tile map for US states
}
\references{
obtained from: \href{https://raw.githubusercontent.com/Z3tt/30DayMapChallenge/master/data/us_states_hexgrid.geojson.json}{Cedric Scherer's Github}
}
