% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.map.lores}
\alias{state.map.lores}
\title{An sf containing lower resolution geometry data for US states}
\usage{
data(state.map.lores)
}
\description{
An sf containing lower resolution geometry data for US states
}
\references{
obtained using tigris::states()
}
