% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{state.names}
\alias{state.names}
\title{A data.frame consisting of all 50 state names plus their postal code abbreviations and FIPS codes.}
\usage{
data(state.names)
}
\description{
choroplethr requires you to use the naming convention in the "region" column (i.e. all lowercase, full name).
}
\references{
Taken from http://www.epa.gov/envirofw/html/codes/state.html
}

