% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{StateChoropleth}
\alias{StateChoropleth}
\title{Create a state-level choropleth}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fa6fd5404e8> 
 - attr(*, "name")= chr "StateChoropleth_generator"
}}
\usage{
StateChoropleth
}
\description{
Create a state-level choropleth
}
\keyword{datasets}

