% Generated by roxygen2 (4.0.0): do not edit by hand
\name{theme_clean}
\alias{theme_clean}
\title{Create a simple ggplot2 theme for rendering choropleths}
\usage{
theme_clean()
}
\description{
Removes axes, margins and sets the background to white.
}
\references{
This code, with minor modifications comes from section 13.19
}

