% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract.R
\name{TractChoropleth}
\alias{TractChoropleth}
\title{An R6 object for creating choropleths of Census Tracts.}
\description{
An R6 object for creating choropleths of Census Tracts.

An R6 object for creating choropleths of Census Tracts.
}
\section{Super class}{
\code{\link[choroplethr:Choropleth]{choroplethr::Choropleth}} -> \code{TractChoropleth}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TractChoropleth-new}{\code{TractChoropleth$new()}}
\item \href{#method-TractChoropleth-set_zoom_tract}{\code{TractChoropleth$set_zoom_tract()}}
\item \href{#method-TractChoropleth-clone}{\code{TractChoropleth$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="bind"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-bind'><code>choroplethr::Choropleth$bind()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="clip"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-clip'><code>choroplethr::Choropleth$clip()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="discretize"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-discretize'><code>choroplethr::Choropleth$discretize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="format_levels"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-format_levels'><code>choroplethr::Choropleth$format_levels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_bounding_box"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_bounding_box'><code>choroplethr::Choropleth$get_bounding_box()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_choropleth_as_polygon"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_choropleth_as_polygon'><code>choroplethr::Choropleth$get_choropleth_as_polygon()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_max_lat"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_max_lat'><code>choroplethr::Choropleth$get_max_lat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_max_long"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_max_long'><code>choroplethr::Choropleth$get_max_long()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_min_lat"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_min_lat'><code>choroplethr::Choropleth$get_min_lat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_min_long"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_min_long'><code>choroplethr::Choropleth$get_min_long()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_reference_map"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_reference_map'><code>choroplethr::Choropleth$get_reference_map()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_scale"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_scale'><code>choroplethr::Choropleth$get_scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_x_scale"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_x_scale'><code>choroplethr::Choropleth$get_x_scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_y_scale"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_y_scale'><code>choroplethr::Choropleth$get_y_scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_zoom"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_zoom'><code>choroplethr::Choropleth$get_zoom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="prepare_map"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-prepare_map'><code>choroplethr::Choropleth$prepare_map()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="render"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-render'><code>choroplethr::Choropleth$render()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="render_with_reference_map"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-render_with_reference_map'><code>choroplethr::Choropleth$render_with_reference_map()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="set_num_colors"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-set_num_colors'><code>choroplethr::Choropleth$set_num_colors()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="set_zoom"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-set_zoom'><code>choroplethr::Choropleth$set_zoom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="theme_clean"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-theme_clean'><code>choroplethr::Choropleth$theme_clean()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="theme_inset"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-theme_inset'><code>choroplethr::Choropleth$theme_inset()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TractChoropleth-new"></a>}}
\if{latex}{\out{\hypertarget{method-TractChoropleth-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TractChoropleth$new(state_name, user.df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TractChoropleth-set_zoom_tract"></a>}}
\if{latex}{\out{\hypertarget{method-TractChoropleth-set_zoom_tract}{}}}
\subsection{Method \code{set_zoom_tract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TractChoropleth$set_zoom_tract(county_zoom, tract_zoom)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TractChoropleth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TractChoropleth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TractChoropleth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
