% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/usa.R
\docType{data}
\name{USAChoropleth}
\alias{USAChoropleth}
\title{Normal choropleth that draws Alaska and Hawaii as insets.
In addition to a columns named "region" and "value", also requires a column named "state".}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x1089d46e0> 
 - attr(*, "name")= chr "USAChoropleth_generator"
}}
\usage{
USAChoropleth
}
\description{
Normal choropleth that draws Alaska and Hawaii as insets.
In addition to a columns named "region" and "value", also requires a column named "state".
}
\keyword{datasets}

