% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zip_choropleth.R
\name{zip_choropleth}
\alias{zip_choropleth}
\title{Create a zip-level choropleth}
\usage{
zip_choropleth(df, title = "", legend = "", buckets = 7, zoom = NULL)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".  Elements in
the "region" column must exactly match how regions are named in the "region" column in ?zip.map.}

\item{title}{An optional title for the map.}

\item{legend}{An optional name for the legend.}

\item{buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.}

\item{zoom}{An optional vector of countries to zoom in on. Elements of this vector must exactly
match the names of countries as they appear in the "region" column of ?zip.regions.}
}
\description{
The map used is zip.map in the choroplethrZip package. See zip.regions for
an object which can help you coerce your regions into the required format.
}

