% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zip_lookup}
\alias{zip_lookup}
\title{A dataset that maps zip codes to counties, states, and CBSAs}
\usage{
data(zip_lookup)
}
\description{
A dataset that maps zip codes to counties, states, and CBSAs
}
\references{
Data obtained from Census Department
}
