% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zip_choropleth.R
\docType{data}
\name{ZipChoropleth}
\alias{ZipChoropleth}
\title{An R6 object for creating ZIP-level choropleths.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x10891f188> 
 - attr(*, "name")= chr "ZipChoropleth_generator"
}}
\usage{
ZipChoropleth
}
\description{
An R6 object for creating ZIP-level choropleths.
}
\keyword{datasets}

