% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zip.R
\docType{data}
\name{ZipMap}
\alias{ZipMap}
\title{Create a county-level choropleth}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x108958f90> 
 - attr(*, "name")= chr "CountyChoropleth_generator"
}}
\usage{
ZipMap
}
\description{
Create a county-level choropleth
}
\keyword{datasets}

