% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin1.R
\name{Admin1Choropleth}
\alias{Admin1Choropleth}
\title{An R6 object for creating Administration Level 1 choropleths.}
\description{
An R6 object for creating Administration Level 1 choropleths.

An R6 object for creating Administration Level 1 choropleths.
}
\section{Super class}{
\code{\link[choroplethr:Choropleth]{choroplethr::Choropleth}} -> \code{Admin1Choropleth}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Admin1Choropleth-new}{\code{Admin1Choropleth$new()}}
\item \href{#method-Admin1Choropleth-clone}{\code{Admin1Choropleth$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_ggscale"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_ggscale'><code>choroplethr::Choropleth$get_ggscale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="guess_geoid_type"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-guess_geoid_type'><code>choroplethr::Choropleth$guess_geoid_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="render"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-render'><code>choroplethr::Choropleth$render()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="set_zoom"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-set_zoom'><code>choroplethr::Choropleth$set_zoom()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Admin1Choropleth-new"></a>}}
\if{latex}{\out{\hypertarget{method-Admin1Choropleth-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Admin1Choropleth$new(country.name, user.df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Admin1Choropleth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Admin1Choropleth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Admin1Choropleth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
