% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/admin1.R
\docType{data}
\name{Admin1Choropleth}
\alias{Admin1Choropleth}
\title{An R6 object for creating Administration Level 1 choropleths.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x10e546ed0> 
 - attr(*, "name")= chr "Admin1Choropleth_generator"
}}
\usage{
Admin1Choropleth
}
\description{
An R6 object for creating Administration Level 1 choropleths.
}
\keyword{datasets}

