% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin1.R
\docType{data}
\name{Admin1Choropleth}
\alias{Admin1Choropleth}
\title{An R6 object for creating Administration Level 1 choropleths.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Admin1Choropleth
}
\description{
An R6 object for creating Administration Level 1 choropleths.
}
\keyword{datasets}
