% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin1_region.R
\docType{data}
\name{Admin1RegionChoropleth}
\alias{Admin1RegionChoropleth}
\title{An R6 object for creating Administration Level 1 choropleths based on regions.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Admin1RegionChoropleth
}
\description{
Compare with the Admin1Choropleth object, which creates Admin 1 choropleths based on 
Countries. This function is useful if you want a map that spans multiple countries -
Especially if it only needs to include a portion of a country.
}
\keyword{datasets}
