% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{Choropleth}
\alias{Choropleth}
\title{The base Choropleth object.}
\description{
The base Choropleth object.

The base Choropleth object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Choropleth$new()}}
\item \href{#method-render}{\code{Choropleth$render()}}
\item \href{#method-get_min_long}{\code{Choropleth$get_min_long()}}
\item \href{#method-get_max_long}{\code{Choropleth$get_max_long()}}
\item \href{#method-get_min_lat}{\code{Choropleth$get_min_lat()}}
\item \href{#method-get_max_lat}{\code{Choropleth$get_max_lat()}}
\item \href{#method-get_bounding_box}{\code{Choropleth$get_bounding_box()}}
\item \href{#method-get_x_scale}{\code{Choropleth$get_x_scale()}}
\item \href{#method-get_y_scale}{\code{Choropleth$get_y_scale()}}
\item \href{#method-get_reference_map}{\code{Choropleth$get_reference_map()}}
\item \href{#method-get_choropleth_as_polygon}{\code{Choropleth$get_choropleth_as_polygon()}}
\item \href{#method-render_with_reference_map}{\code{Choropleth$render_with_reference_map()}}
\item \href{#method-clip}{\code{Choropleth$clip()}}
\item \href{#method-discretize}{\code{Choropleth$discretize()}}
\item \href{#method-bind}{\code{Choropleth$bind()}}
\item \href{#method-prepare_map}{\code{Choropleth$prepare_map()}}
\item \href{#method-get_scale}{\code{Choropleth$get_scale()}}
\item \href{#method-theme_clean}{\code{Choropleth$theme_clean()}}
\item \href{#method-theme_inset}{\code{Choropleth$theme_inset()}}
\item \href{#method-format_levels}{\code{Choropleth$format_levels()}}
\item \href{#method-set_zoom}{\code{Choropleth$set_zoom()}}
\item \href{#method-get_zoom}{\code{Choropleth$get_zoom()}}
\item \href{#method-set_num_colors}{\code{Choropleth$set_num_colors()}}
\item \href{#method-clone}{\code{Choropleth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$new(map.df, user.df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render"></a>}}
\if{latex}{\out{\hypertarget{method-render}{}}}
\subsection{Method \code{render()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$render()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_min_long"></a>}}
\if{latex}{\out{\hypertarget{method-get_min_long}{}}}
\subsection{Method \code{get_min_long()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$get_min_long()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_max_long"></a>}}
\if{latex}{\out{\hypertarget{method-get_max_long}{}}}
\subsection{Method \code{get_max_long()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$get_max_long()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_min_lat"></a>}}
\if{latex}{\out{\hypertarget{method-get_min_lat}{}}}
\subsection{Method \code{get_min_lat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$get_min_lat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_max_lat"></a>}}
\if{latex}{\out{\hypertarget{method-get_max_lat}{}}}
\subsection{Method \code{get_max_lat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$get_max_lat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_bounding_box"></a>}}
\if{latex}{\out{\hypertarget{method-get_bounding_box}{}}}
\subsection{Method \code{get_bounding_box()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$get_bounding_box(long_margin_percent, lat_margin_percent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_x_scale"></a>}}
\if{latex}{\out{\hypertarget{method-get_x_scale}{}}}
\subsection{Method \code{get_x_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$get_x_scale()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_y_scale"></a>}}
\if{latex}{\out{\hypertarget{method-get_y_scale}{}}}
\subsection{Method \code{get_y_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$get_y_scale()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_reference_map"></a>}}
\if{latex}{\out{\hypertarget{method-get_reference_map}{}}}
\subsection{Method \code{get_reference_map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$get_reference_map()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_choropleth_as_polygon"></a>}}
\if{latex}{\out{\hypertarget{method-get_choropleth_as_polygon}{}}}
\subsection{Method \code{get_choropleth_as_polygon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$get_choropleth_as_polygon(alpha)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_with_reference_map"></a>}}
\if{latex}{\out{\hypertarget{method-render_with_reference_map}{}}}
\subsection{Method \code{render_with_reference_map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$render_with_reference_map(alpha = 0.5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clip"></a>}}
\if{latex}{\out{\hypertarget{method-clip}{}}}
\subsection{Method \code{clip()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$clip()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-discretize"></a>}}
\if{latex}{\out{\hypertarget{method-discretize}{}}}
\subsection{Method \code{discretize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$discretize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bind"></a>}}
\if{latex}{\out{\hypertarget{method-bind}{}}}
\subsection{Method \code{bind()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$bind()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare_map"></a>}}
\if{latex}{\out{\hypertarget{method-prepare_map}{}}}
\subsection{Method \code{prepare_map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$prepare_map()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_scale"></a>}}
\if{latex}{\out{\hypertarget{method-get_scale}{}}}
\subsection{Method \code{get_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$get_scale()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-theme_clean"></a>}}
\if{latex}{\out{\hypertarget{method-theme_clean}{}}}
\subsection{Method \code{theme_clean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$theme_clean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-theme_inset"></a>}}
\if{latex}{\out{\hypertarget{method-theme_inset}{}}}
\subsection{Method \code{theme_inset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$theme_inset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_levels"></a>}}
\if{latex}{\out{\hypertarget{method-format_levels}{}}}
\subsection{Method \code{format_levels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$format_levels(x, nsep = " to ")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_zoom"></a>}}
\if{latex}{\out{\hypertarget{method-set_zoom}{}}}
\subsection{Method \code{set_zoom()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$set_zoom(zoom)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_zoom"></a>}}
\if{latex}{\out{\hypertarget{method-get_zoom}{}}}
\subsection{Method \code{get_zoom()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$get_zoom()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_num_colors"></a>}}
\if{latex}{\out{\hypertarget{method-set_num_colors}{}}}
\subsection{Method \code{set_num_colors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$set_num_colors(num_colors)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
