% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{Choropleth}
\alias{Choropleth}
\title{The base Choropleth object.}
\description{
The base Choropleth object.

The base Choropleth object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Choropleth-new}{\code{Choropleth$new()}}
\item \href{#method-Choropleth-guess_geoid_type}{\code{Choropleth$guess_geoid_type()}}
\item \href{#method-Choropleth-set_zoom}{\code{Choropleth$set_zoom()}}
\item \href{#method-Choropleth-get_ggscale}{\code{Choropleth$get_ggscale()}}
\item \href{#method-Choropleth-render}{\code{Choropleth$render()}}
\item \href{#method-Choropleth-clone}{\code{Choropleth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth-new"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$new(user.df, geoid.name, geoid.type, value.name, num_colors)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth-guess_geoid_type"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth-guess_geoid_type}{}}}
\subsection{Method \code{guess_geoid_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$guess_geoid_type(user.regions)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth-set_zoom"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth-set_zoom}{}}}
\subsection{Method \code{set_zoom()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$set_zoom(zoom)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth-get_ggscale"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth-get_ggscale}{}}}
\subsection{Method \code{get_ggscale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$get_ggscale(
  choropleth.df = self$choropleth.df,
  custom.colors,
  color.min,
  color.max,
  na.color,
  nbreaks
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth-render"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth-render}{}}}
\subsection{Method \code{render()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$render(
  choropleth.df = self$choropleth.df,
  ggscale,
  projection = "cartesian",
  limits_lat = NULL,
  limits_lon = NULL,
  reproject = TRUE,
  occlude_latlon_limits = TRUE,
  ignore_latlon = FALSE,
  border_color = "black",
  border_thickness = 0.2,
  background_color = "white",
  gridlines = FALSE,
  latlon_ticks = FALSE,
  label = NULL,
  label_text_size,
  label_text_color,
  label_box_color,
  ggrepel_options = NULL,
  legend = NULL,
  legend_position = "right",
  title = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
