% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/choropleth.R
\docType{data}
\name{Choropleth}
\alias{Choropleth}
\title{The base Choropleth object.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x113522e08> 
 - attr(*, "name")= chr "Choropleth_generator"
}}
\usage{
Choropleth
}
\description{
The base Choropleth object.
}
\keyword{datasets}

