% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\docType{data}
\name{Choropleth}
\alias{Choropleth}
\title{The base Choropleth object.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Choropleth
}
\description{
The base Choropleth object.
}
\keyword{datasets}
