% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{choroplethr}
\alias{choroplethr}
\title{Create a choropleth}
\usage{
choroplethr(...)
}
\arguments{
\item{...}{All arguments are ignored.}
}
\description{
This function is deprecated as of choroplethr version 2.0.0. Please use ?state_choropleth, 
?county_choropleth, ?zip_map and ?country_choroplethr instead. The last version of choroplethr 
in which this function worked was version 1.7.0, which can be downloaded from CRAN 
here: http://cran.r-project.org/web/packages/choroplethr/index.html
}
