% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country.map}
\alias{country.map}
\title{An sf containing geometry data for countries the world}
\usage{
data(country.map)
}
\description{
An sf containing geometry data for countries the world
}
\references{
obtained using rnatrualearth::ne_countries
}
