% Generated by roxygen2 (4.0.2): do not edit by hand
\name{country_choropleth}
\alias{country_choropleth}
\title{Create a country-level choropleth}
\usage{
country_choropleth(df, title = "", legend_name = "", num_buckets = 7,
  countries = NULL)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".}

\item{title}{An optional title for the map.}

\item{legend_name}{An optional name for the legend.}

\item{num_buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.}

\item{countries}{Which countries of the map to show. If NULL, show all.
Type "data(country.names); ?country.names" to see the list of countries. Countries must
correspond to the "region" column of country.names.}
}
\description{
Create a country-level choropleth
}
\examples{
data(df_pop_country)
country_choropleth(df_pop_country, "2012 World Bank Populate Estimates")
}

