% Generated by roxygen2 (4.0.1): do not edit by hand
\name{country_choropleth}
\alias{country_choropleth}
\title{Create a country-level choropleth}
\usage{
country_choropleth(df, title = "", legend_name = "", num_buckets = 7)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".}

\item{title}{An optional title for the map.}

\item{legend_name}{An optional name for the legend.}

\item{num_buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.}
}
\description{
Create a country-level choropleth
}
\examples{
data(country.names)
data(country.map)
df = data.frame(region=country.names$region, value=sample(1:nrow(country.names)))
country_choropleth(df)
}

