% Generated by roxygen2 (4.0.1): do not edit by hand
\name{country_choropleth}
\alias{country_choropleth}
\title{Create a choropleth of most countries of the world}
\usage{
country_choropleth(df, title = "", scaleName = "", num_buckets = 7,
  warn_na = FALSE, countries = NULL)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".  Region must contain
county names as defined in the region column of \code{\link{country.names}}.}

\item{title}{An optional title for the map.}

\item{scaleName}{An optional label for the legend.}

\item{num_buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.  For
example, 2 will show values above or below the median, and 9 will show the maximum
resolution.}

\item{warn_na}{If true, choroplethr will emit a warning when a) you give it regions that it is ignoring and b) you do not supply regions that it is rendering.}

\item{countries}{A vector of countries to render.}
}
\value{
A choropleth.
}
\description{
Create a choropleth of most countries of the world
}
\examples{
data(country.names)

df = data.frame(region=country.names$region, value=sample.int(nrow(country.names)))
country_choropleth(df)
}
\seealso{
\code{\link{country.map}} which contains information about the country map and \code{\link{country.names}} which contains the names of the regions in the map.
}
\keyword{choropleth}

