% Generated by roxygen2 (4.0.2): do not edit by hand
\name{country_choropleth}
\alias{country_choropleth}
\title{Create a country-level choropleth}
\usage{
country_choropleth(df, title = "", legend = "", buckets = 7,
  zoom = NULL)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".  Elements in
the "region" column must exactly match how regions are named in the "region" column in ?country.map.
See ?country.names for an object which can help you coerce your regions into the required format.}

\item{title}{An optional title for the map.}

\item{legend}{An optional name for the legend.}

\item{buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.}

\item{zoom}{An optional vector of countries to zoom in on. Elements of this vector must exactly
match the names of countries as they appear in the "region" column of ?country.names}
}
\description{
The map used is ?country.map.
}
\examples{
# demonstrate default options
data(df_pop_country)
country_choropleth(df_pop_country, "2012 World Bank Populate Estimates")

# demonstrate continuous scale
country_choropleth(df_pop_country, "2012 World Bank Populate Estimates", buckets=1)

# demonstrate zooming
country_choropleth(df_pop_country,
                   "2012 World Bank Population Estimates",
                   buckets=1,
                   zoom=c("united states of america", "canada", "mexico"))
}

