% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{CountryChoropleth}
\alias{CountryChoropleth}
\title{An R6 object for creating country-level choropleths.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7f9f7f9d84b0> 
 - attr(*, "name")= chr "CountryChoropleth_generator"
}}
\usage{
CountryChoropleth
}
\description{
An R6 object for creating country-level choropleths.
}
\keyword{datasets}

