% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/country.R
\docType{data}
\name{CountryChoropleth}
\alias{CountryChoropleth}
\title{An R6 object for creating country-level choropleths.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x10260f0d0> 
 - attr(*, "name")= chr "CountryChoropleth_generator"
}}
\usage{
CountryChoropleth
}
\description{
An R6 object for creating country-level choropleths.
}
\keyword{datasets}

