% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.R
\docType{data}
\name{CountryChoropleth}
\alias{CountryChoropleth}
\title{An R6 object for creating country-level choropleths.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CountryChoropleth
}
\description{
An R6 object for creating country-level choropleths.
}
\keyword{datasets}
