% Generated by roxygen2 (4.0.1): do not edit by hand
\name{county_choropleth}
\alias{county_choropleth}
\title{Create a choropleth of US Counties}
\usage{
county_choropleth(df, title = "", scaleName = "", num_buckets = 7,
  warn_na = FALSE, states = state.abb, renderAsInsets = TRUE,
  scale = NULL)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".  Region must contain county FIPS codes.}

\item{title}{An optional title for the map.}

\item{scaleName}{An optional label for the legend.}

\item{num_buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.  For
example, 2 will show values above or below the median, and 9 will show the maximum
resolution.}

\item{warn_na}{If true, choroplethr will emit a warning when a) you give it regions that it is ignoring and b) you do not supply regions that it is rendering.}

\item{states}{A vector of states to render.  Defaults to state.abb.}

\item{renderAsInsets}{If true, Alaska and Hawaii will be rendered as insets on the map.  If false, all 50 states will be rendered
on the same longitude and latitude map to scale. This variable is only checked when the "states" variable is equal to all 50 states.}

\item{scale}{an optional ggplot2 scale to use. For example "scale_fill_brewer(palette=2)".}
}
\value{
A choropleth.
}
\description{
Create a choropleth of US Counties
}
\examples{
data(choroplethr)

# 2012 county population estimates - continuous vs. discrete scale
county_choropleth(df_pop_county, num_buckets=1, title="2012 County Population Estimates")
county_choropleth(df_pop_county, num_buckets=9, title="2012 County Population Estimates")
}
\seealso{
\code{\link{county.map}} which contains information about the county map and \code{\link{county.names}} which contains the names of the regions in the map.
}
\keyword{choropleth}

