% Generated by roxygen2 (4.0.1): do not edit by hand
\name{county_choropleth}
\alias{county_choropleth}
\title{Create a choropleth of USA Counties with sensible defaults.}
\usage{
county_choropleth(df, title = "", legend_name = "", num_buckets = 7)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".}

\item{title}{An optional title for the map.}

\item{legend_name}{An optional name for the legend.}

\item{num_buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.}
}
\description{
Create a choropleth of USA Counties with sensible defaults.
}

