% Generated by roxygen2 (4.0.2): do not edit by hand
\name{county_choropleth}
\alias{county_choropleth}
\title{Create a choropleth of USA Counties with sensible defaults.}
\usage{
county_choropleth(df, title = "", legend_name = "", num_buckets = 7,
  zoom = NULL)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".}

\item{title}{An optional title for the map.}

\item{legend_name}{An optional name for the legend.}

\item{num_buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.}

\item{zoom}{An optional list of states to zoom in on. Must come from the "name" column in
?state.names.}
}
\description{
Create a choropleth of USA Counties with sensible defaults.
}
\examples{
data(df_pop_county)
county_choropleth(df_pop_county, title="US 2012 Population Estimates", legend_name="Population")
}

