% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acs.R
\name{county_choropleth_acs}
\alias{county_choropleth_acs}
\title{Create a US County choropleth from ACS data.}
\usage{
county_choropleth_acs(tableId, endyear = 2011, span = 5, buckets = 7,
  zoom = NULL)
}
\arguments{
\item{tableId}{The id of an ACS table}

\item{endyear}{The end year of the survey to use.  See acs.fetch (?acs.fetch) and http://1.usa.gov/1geFSSj for details.}

\item{span}{The span of time to use.  See acs.fetch and http://1.usa.gov/1geFSSj for details.}

\item{buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.}

\item{zoom}{An optional list of states to zoom in on. Must come from the "name" column in
?state.regions.}
}
\value{
A choropleth.
}
\description{
Creates a US County choropleth using the US Census' American Community Survey (ACS) data.
Requires the acs package to be installed, and a Census API Key to be set with
the acs's api.key.install function.  Census API keys can be obtained at http://www.census.gov/developers/tos/key_request.html.
}
\examples{
\dontrun{
# population of all counties
county_choropleth_acs("B01003")

# median income, continuous scale, counties in New York, New Jersey and Connecticut
county_choropleth_acs("B19301", buckets=1, zoom=c("new york", "new jersey", "connecticut"))
}
}
\references{
Uses the acs package created by Ezra Haber Glenn.
}
\seealso{
\code{api.key.install} in the acs package which sets an Census API key for the acs library

http://factfinder2.census.gov/faces/help/jsf/pages/metadata.xhtml?lang=en&type=survey&id=survey.en.ACS_ACS
which contains a list of all ACS surveys.
}
\keyword{acs}
\keyword{choropleth,}

