% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{county_zoom_choropleth}
\alias{county_zoom_choropleth}
\title{Create a choropleth of USA Counties, with sensible defaults, that zooms on counties.}
\usage{
county_zoom_choropleth(...)
}
\arguments{
\item{...}{All arguments are ignored.}
}
\description{
This function is deprecated as of choroplethr version 3.0.0. Please use ?county_choropleth
with the county_zoom parameter set instead.  The last version of choroplethr 
in which this function worked was version 2.1.1, which can be downloaded from CRAN 
here: http://cran.r-project.org/web/packages/choroplethr/index.html
}
