% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county.R
\name{CountyChoropleth}
\alias{CountyChoropleth}
\title{Create a county-level choropleth}
\description{
Create a county-level choropleth

Create a county-level choropleth
}
\section{Super classes}{
\code{\link[choroplethr:Choropleth]{choroplethr::Choropleth}} -> \code{\link[choroplethr:USAChoropleth]{choroplethr::USAChoropleth}} -> \code{CountyChoropleth}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CountyChoropleth-new}{\code{CountyChoropleth$new()}}
\item \href{#method-CountyChoropleth-clip}{\code{CountyChoropleth$clip()}}
\item \href{#method-CountyChoropleth-clone}{\code{CountyChoropleth$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_ggscale"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_ggscale'><code>choroplethr::Choropleth$get_ggscale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="guess_geoid_type"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-guess_geoid_type'><code>choroplethr::Choropleth$guess_geoid_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="USAChoropleth" data-id="render"><a href='../../choroplethr/html/USAChoropleth.html#method-USAChoropleth-render'><code>choroplethr::USAChoropleth$render()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="USAChoropleth" data-id="render_helper"><a href='../../choroplethr/html/USAChoropleth.html#method-USAChoropleth-render_helper'><code>choroplethr::USAChoropleth$render_helper()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="USAChoropleth" data-id="render_state_outline"><a href='../../choroplethr/html/USAChoropleth.html#method-USAChoropleth-render_state_outline'><code>choroplethr::USAChoropleth$render_state_outline()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="USAChoropleth" data-id="set_zoom"><a href='../../choroplethr/html/USAChoropleth.html#method-USAChoropleth-set_zoom'><code>choroplethr::USAChoropleth$set_zoom()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CountyChoropleth-new"></a>}}
\if{latex}{\out{\hypertarget{method-CountyChoropleth-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountyChoropleth$new(user.df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CountyChoropleth-clip"></a>}}
\if{latex}{\out{\hypertarget{method-CountyChoropleth-clip}{}}}
\subsection{Method \code{clip()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountyChoropleth$clip()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CountyChoropleth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CountyChoropleth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountyChoropleth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
