% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/county.R
\docType{data}
\name{CountyChoropleth}
\alias{CountyChoropleth}
\title{Create a county-level choropleth}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x112162030> 
 - attr(*, "name")= chr "CountyChoropleth_generator"
}}
\usage{
CountyChoropleth
}
\description{
Create a county-level choropleth
}
\keyword{datasets}

