% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{CountyChoropleth}
\alias{CountyChoropleth}
\title{Create a county-level choropleth}
\format{\preformatted{List of 8
 $ new       :function (...)  
 $ classname : chr "CountyChoropleth"
 $ public    :List of 21
  ..$ title        : chr ""
  ..$ legend_name  : chr ""
  ..$ warn         : logi TRUE
  ..$ ggplot_scale : NULL
  ..$ user.df      : NULL
  ..$ map.df       : NULL
  ..$ choropleth.df: NULL
  ..$ map.names    : NULL
  ..$ num_buckets  : num 7
  ..$ regions      : NULL
  ..$ clip         :function ()  
  ..$ discretize   :function ()  
  ..$ bind         :function ()  
  ..$ prepare_map  :function ()  
  ..$ get_scale    :function (min = NA, max = NA)  
  ..$ theme_clean  :function ()  
  ..$ theme_inset  :function ()  
  ..$ format_levels:function (x, nsep = " to ")  
  ..$ initialize   :function (user.df)  
  ..$ render       :function (num_buckets = 7)  
  ..$ render_helper:function (choropleth.df, scale_name, theme, min, max)  
 $ private   : NULL
 $ active    : NULL
 $ inherit   :List of 8
  ..$ new       :function (...)  
  ..$ classname : chr "Choropleth"
  ..$ public    :List of 20
  .. ..$ title        : chr ""
  .. ..$ legend_name  : chr ""
  .. ..$ warn         : logi TRUE
  .. ..$ ggplot_scale : NULL
  .. ..$ initialize   :function (map.df, user.df)  
  .. ..$ render       :function (num_buckets = 7)  
  .. ..$ user.df      : NULL
  .. ..$ map.df       : NULL
  .. ..$ choropleth.df: NULL
  .. ..$ map.names    : NULL
  .. ..$ num_buckets  : num 7
  .. ..$ regions      : NULL
  .. ..$ clip         :function ()  
  .. ..$ discretize   :function ()  
  .. ..$ bind         :function ()  
  .. ..$ prepare_map  :function ()  
  .. ..$ get_scale    :function (min = NA, max = NA)  
  .. ..$ theme_clean  :function ()  
  .. ..$ theme_inset  :function ()  
  .. ..$ format_levels:function (x, nsep = " to ")  
  ..$ private   : NULL
  ..$ active    : NULL
  ..$ inherit   : NULL
  ..$ parent_env:<environment: 0x7f90beb11430> 
  ..$ lock      : logi TRUE
  ..- attr(*, "class")= chr "R6ClassGenerator"
 $ parent_env:<environment: 0x7f90beb11430> 
 $ lock      : logi TRUE
 - attr(*, "class")= chr "R6ClassGenerator"
}}
\usage{
CountyChoropleth
}
\description{
Create a county-level choropleth
}
\keyword{datasets}

