% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{CountyChoropleth}
\alias{CountyChoropleth}
\title{Create a county-level choropleth}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fbc841812d0> 
 - attr(*, "name")= chr "CountyChoropleth_generator"
}}
\usage{
CountyChoropleth
}
\description{
Create a county-level choropleth
}
\keyword{datasets}

