% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county.R
\docType{data}
\name{CountyChoropleth}
\alias{CountyChoropleth}
\title{Create a county-level choropleth}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CountyChoropleth
}
\description{
Create a county-level choropleth
}
\keyword{datasets}
