% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_zoom.R
\name{CountyZoomChoropleth}
\alias{CountyZoomChoropleth}
\title{Create a county-level choropleth that zooms on counties, not states.}
\description{
Create a county-level choropleth that zooms on counties, not states.

Create a county-level choropleth that zooms on counties, not states.
}
\section{Super class}{
\code{\link[choroplethr:Choropleth]{choroplethr::Choropleth}} -> \code{CountyZoomChoropleth}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CountyZoomChoropleth-new}{\code{CountyZoomChoropleth$new()}}
\item \href{#method-CountyZoomChoropleth-render}{\code{CountyZoomChoropleth$render()}}
\item \href{#method-CountyZoomChoropleth-clone}{\code{CountyZoomChoropleth$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_ggscale"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_ggscale'><code>choroplethr::Choropleth$get_ggscale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="guess_geoid_type"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-guess_geoid_type'><code>choroplethr::Choropleth$guess_geoid_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="set_zoom"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-set_zoom'><code>choroplethr::Choropleth$set_zoom()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CountyZoomChoropleth-new"></a>}}
\if{latex}{\out{\hypertarget{method-CountyZoomChoropleth-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountyZoomChoropleth$new(user.df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CountyZoomChoropleth-render"></a>}}
\if{latex}{\out{\hypertarget{method-CountyZoomChoropleth-render}{}}}
\subsection{Method \code{render()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountyZoomChoropleth$render()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CountyZoomChoropleth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CountyZoomChoropleth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountyZoomChoropleth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
