% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{CountyZoomChoropleth}
\alias{CountyZoomChoropleth}
\title{Create a county-level choropleth that zooms on counties, not states.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7f9d931b4030> 
 - attr(*, "name")= chr "CountyZoomChoropleth_generator"
}}
\usage{
CountyZoomChoropleth
}
\description{
Create a county-level choropleth that zooms on counties, not states.
}
\keyword{datasets}

