% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/county_zoom.R
\docType{data}
\name{CountyZoomChoropleth}
\alias{CountyZoomChoropleth}
\title{Create a county-level choropleth that zooms on counties, not states.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x10340ad08> 
 - attr(*, "name")= chr "CountyZoomChoropleth_generator"
}}
\usage{
CountyZoomChoropleth
}
\description{
Create a county-level choropleth that zooms on counties, not states.
}
\keyword{datasets}

