% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_zoom.R
\docType{data}
\name{CountyZoomChoropleth}
\alias{CountyZoomChoropleth}
\title{Create a county-level choropleth that zooms on counties, not states.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CountyZoomChoropleth
}
\description{
Create a county-level choropleth that zooms on counties, not states.
}
\keyword{datasets}
