% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_japan_census}
\alias{df_japan_census}
\title{A data.frame containing basic demographic information about Japan.}
\usage{
data(df_japan_census)
}
\description{
A data.frame containing basic demographic information about Japan.
}
\references{
Taken from the "Total Population" table from the Statistics Bureau of Japan website
(\url{http://www.stat.go.jp/english/data/nenkan/1431-02.htm}) on 12/1/2014.
}
\keyword{data}
