% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pop_country}
\alias{df_pop_country}
\title{A data.frame containing population estimates for Countries in 2012.}
\usage{
data(df_pop_country)
}
\description{
A data.frame containing population estimates for Countries in 2012.
}
\references{
Taken from the WDI package with code SP.POP.TOTL for year 2012.
}
\keyword{data}
