% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{df_pop_county}
\alias{df_pop_county}
\title{A data.frame containing population estimates for US Counties in 2012.}
\usage{
data(choroplethr)
}
\description{
A data.frame containing population estimates for US Counties in 2012.
}
\examples{
data(choroplethr)
choroplethr(df_pop_state, "state", title="2012 Population Estimates")
}
\references{
Taken from the US American Community Survey (ACS) 5 year estimates with \link{get_acs_df} using table B01003.
}
\keyword{data}

