% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pop_county}
\alias{df_pop_county}
\title{A data.frame containing population estimates for US Counties in 2012.}
\usage{
data(df_pop_county)
}
\description{
A data.frame containing population estimates for US Counties in 2012.
}
\references{
Taken from the US American Community Survey (ACS) 5 year estimates.
}
\keyword{data}
