% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{df_pop_zip}
\alias{df_pop_zip}
\title{A data.frame containing population estimates for US Zip Code Tabulated Areas (ZCTAs) in 2012.}
\description{
ZCTAs are intended to be roughly analogous to postal ZIP codes.
}
\examples{
data(choroplethr)
choroplethr(df_pop_state, "state", title="2012 Population Estimates")
}
\references{
Taken from the US American Community Survey (ACS) 5 year estimates with \link{get_acs_df} using table B01003.
ZCTAs, and their realationship to ZIP codes, are explained here \url{https://www.census.gov/geo/reference/zctas.html}.
}
\keyword{data}

