% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{df_president}
\alias{df_president}
\title{A data.frame containing election results from the 2012 US Presidential election.}
\usage{
data(choroplethr)
}
\description{
A data.frame containing election results from the 2012 US Presidential election.
}
\references{
Taken from \url{http://www.fec.gov/pubrec/fe2012/federalelections2012.shtml} on 3/15/2014.
}
\keyword{data}

