% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_acs_df}
\alias{get_acs_df}
\title{Returns a data.frame representing US Census American Community Survey (ACS) estimates.}
\usage{
get_acs_df(tableId, lod, endyear = 2012, span = 5, column_idx = -1)
}
\arguments{
\item{tableId}{The id of an ACS table.}

\item{lod}{The level of geographic detail for the data.frame.  Must be one of "state", "county" or "zip".}

\item{endyear}{The end year of the survey.  Defaults to 2012.}

\item{span}{The span of the survey.  Defaults to 5.}

\item{column_idx}{An optional column index to specify.}
}
\value{
A data.frame.
}
\description{
For a walkthru on the ACS, see the vignette "Mapping Census Data".
Requires the acs package to be installed, and a Census API Key to be set with the
acs's api.key.install function.  Census API keys can be obtained at http://www.census.gov/developers/tos/key_request.html.
The resulting data.frame is in a format which can be accepted by the functions state_choropleth,
county_choropleth and zip_choropleth.
}
\examples{
\dontrun{

# some simple examples - that table B01003 is population
state_choropleth(get_acs_df("B01003", "state"))
county_choropleth(get_acs_df("B01003", "county"))

# compare states with >1M people with counties with >1M people
library(Hmisc) # for cut2
df = get_acs_df("B01003", "state")
df$value = cut2(df$value,cuts=c(0,1000000,Inf))
state_choropleth(df, "States with a population over 1M", "Population")

df = get_acs_df("B01003", "county")
df$value = cut2(df$value, cuts=c(0,1000000,Inf))
county_choropleth(df, "Counties with a population over 1M", "Population")

# ZIP codes in California where median age is between 20 and 30
df       = get_acs_df("B01002", "zip") # median age
df       = df[df$value >= 20 & df$value <= 30, ]
df$value = cut2(df$value, g=3) # 3 equally-sized groups
zip_choropleth(df,
               title = "CA Zip Codes by Age",
              scaleName = "Median Age",
              states = "CA")
}
}
\seealso{
http://factfinder2.census.gov/faces/help/jsf/pages/metadata.xhtml?lang=en&type=survey&id=survey.en.ACS_ACS, which lists all ACS Surveys.
}

