% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_acs_df}
\alias{get_acs_df}
\title{Returns a data.frame representing US Census American Community Survey (ACS) estimates.}
\usage{
get_acs_df(...)
}
\arguments{
\item{...}{All arguments are ignored.}
}
\description{
This function is deprecated as of choroplethr version 3.0.0. Please use ?get_acs_data instead.
The last version of choroplethr in which this functions worked was version 2.1.1, which can
be downloaded from CRAN here: http://cran.r-project.org/web/packages/choroplethr/index.html
}
