% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tract_demographics.R
\name{get_tract_demographics}
\alias{get_tract_demographics}
\title{Get a handful of demographic variables on Census Tracts in a State from the US Census Bureau as a data.frame.}
\usage{
get_tract_demographics(state_fips, endyear = 2013, span = 5)
}
\arguments{
\item{state_fips}{The FIPS code of the state you want demographics for}

\item{endyear}{The end year for the survey}

\item{span}{The span of the survey}
}
\description{
The data comes from the American Community Survey (ACS). The variables are: total population, percent White 
not Hispanic, Percent Black or African American not Hispanic, percent Asian not Hispanic,
percent Hispanic all races, per-capita income, median rent and median age.
}
\references{
The choroplethr guide to Census data: http://www.arilamstein.com/open-source/choroplethr/mapping-us-census-data/

A list of all ACS Surveys: http://factfinder.census.gov/faces/affhelp/jsf/pages/metadata.xhtml?lang=en&type=survey&id=survey.en.ACS_ACS
}
