% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts.R
\name{get_tract_map}
\alias{get_tract_map}
\title{Get a map of tracts in a state, as a data.frame}
\usage{
get_tract_map(state_fips)
}
\arguments{
\item{state_fips}{The fips code of the state}
}
\description{
The map returned is exactly the same map which tract_choropleth uses. It is downloaded
using the "tracts" function in the tigris package, and then it is modified for use with 
choroplethr.
}
