% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.map.hires}
\alias{state.map.hires}
\title{An sf containing higher resolution geometry data for US states}
\usage{
data(state.map.hires)
}
\description{
Note: Resolution is still much lower than raw data from tigris
}
\references{
obtained using tigris::states()
}
