% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.map}
\alias{state.map}
\title{An sf containing geometry data for US states}
\usage{
data(state.map)
}
\description{
An sf containing geometry data for US states
}
\references{
obtained using tigris::states()
}
