% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{state.names}
\alias{state.names}
\title{A data.frame consisting of all 50 state names (plus the District of Columbia), postal codes,
and FIPS codes both as characters and numbers (i.e. with and without a leading 0).}
\description{
A data.frame consisting of all 50 state names (plus the District of Columbia), postal codes,
and FIPS codes both as characters and numbers (i.e. with and without a leading 0).
}
\references{
Taken from http://www.epa.gov/envirofw/html/codes/state.html
}

