% Generated by roxygen2 (4.0.1): do not edit by hand
\name{state_choropleth}
\alias{state_choropleth}
\title{Create a choropleth of US States}
\usage{
state_choropleth(df, title = "", scaleName = "", num_buckets = 7,
  warn_na = FALSE, states = state.abb, showLabels = TRUE,
  renderAsInsets = TRUE, scale = NULL)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".  Region must contain state names or abbreviations (e.g. "California" or "CA").}

\item{title}{An optional title for the map.}

\item{scaleName}{An optional label for the legend.}

\item{num_buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.  For
example, 2 will show values above or below the median, and 9 will show the maximum
resolution.}

\item{warn_na}{If true, choroplethr will emit a warning when a) you give it regions that it is ignoring and b) you do not supply regions that it is rendering.}

\item{states}{A vector of states to render.  Defaults to state.abb.}

\item{showLabels}{For state choropleths, whether or not to show state abbreviations on the map.
Defaults to T.}

\item{renderAsInsets}{If true, Alaska and Hawaii will be rendered as insets on the map.  If false, all 50 states will be rendered
on the same longitude and latitude map to scale. This variable is only checked when the "states" variable is equal to all 50 states.}
}
\value{
A choropleth.
}
\description{
Create a choropleth of US States
}
\examples{
data(choroplethr)

# 2012 US Presidential results
state_choropleth(df_president, title="2012 US Presidential Results")

# 2012 state population estimates - continuous vs. discrete scale
state_choropleth(df_pop_state, num_buckets=1, title="2012 State Population Estimates")
state_choropleth(df_pop_state, num_buckets=9, title="2012 State Population Estimates")
}
\seealso{
\code{\link{state.map}} which contains information about the state map and \code{\link{state.names}} which contains the names of the regions in the map.
}
\keyword{choropleth}

