% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{StateChoropleth}
\alias{StateChoropleth}
\title{Create a state-level choropleth}
\description{
Create a state-level choropleth

Create a state-level choropleth
}
\section{Super class}{
\code{\link[choroplethr:Choropleth]{choroplethr::Choropleth}} -> \code{StateChoropleth}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StateChoropleth-clone}{\code{StateChoropleth$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_ggscale"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_ggscale'><code>choroplethr::Choropleth$get_ggscale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="guess_geoid_type"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-guess_geoid_type'><code>choroplethr::Choropleth$guess_geoid_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="initialize"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-initialize'><code>choroplethr::Choropleth$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="render"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-render'><code>choroplethr::Choropleth$render()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="set_zoom"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-set_zoom'><code>choroplethr::Choropleth$set_zoom()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateChoropleth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StateChoropleth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateChoropleth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
