% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/state.R
\docType{data}
\name{StateChoropleth}
\alias{StateChoropleth}
\title{Create a state-level choropleth}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x103c40550> 
 - attr(*, "name")= chr "StateChoropleth_generator"
}}
\usage{
StateChoropleth
}
\description{
Create a state-level choropleth
}
\keyword{datasets}

