% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\docType{data}
\name{StateChoropleth}
\alias{StateChoropleth}
\title{Create a state-level choropleth}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StateChoropleth
}
\description{
Create a state-level choropleth
}
\keyword{datasets}
