% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract.R
\name{TractChoropleth}
\alias{TractChoropleth}
\title{An R6 object for creating choropleths of Census Tracts.}
\description{
An R6 object for creating choropleths of Census Tracts.

An R6 object for creating choropleths of Census Tracts.
}
\section{Super class}{
\code{\link[choroplethr:Choropleth]{choroplethr::Choropleth}} -> \code{TractChoropleth}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TractChoropleth-new}{\code{TractChoropleth$new()}}
\item \href{#method-TractChoropleth-set_zoom_tract}{\code{TractChoropleth$set_zoom_tract()}}
\item \href{#method-TractChoropleth-clone}{\code{TractChoropleth$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_ggscale"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_ggscale'><code>choroplethr::Choropleth$get_ggscale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="guess_geoid_type"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-guess_geoid_type'><code>choroplethr::Choropleth$guess_geoid_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="render"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-render'><code>choroplethr::Choropleth$render()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="set_zoom"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-set_zoom'><code>choroplethr::Choropleth$set_zoom()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TractChoropleth-new"></a>}}
\if{latex}{\out{\hypertarget{method-TractChoropleth-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TractChoropleth$new(state_name, user.df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TractChoropleth-set_zoom_tract"></a>}}
\if{latex}{\out{\hypertarget{method-TractChoropleth-set_zoom_tract}{}}}
\subsection{Method \code{set_zoom_tract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TractChoropleth$set_zoom_tract(county_zoom, tract_zoom)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TractChoropleth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TractChoropleth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TractChoropleth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
