% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts.R
\docType{data}
\name{TractChoropleth}
\alias{TractChoropleth}
\title{An R6 object for creating choropleths of Census Tracts.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TractChoropleth
}
\description{
An R6 object for creating choropleths of Census Tracts.
}
\keyword{datasets}
