% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa.R
\name{USAChoropleth}
\alias{USAChoropleth}
\title{Normal choropleth that draws Alaska and Hawaii as insets. 
In addition to a columns named "region" and "value", also requires a column named "state".}
\description{
Normal choropleth that draws Alaska and Hawaii as insets. 
In addition to a columns named "region" and "value", also requires a column named "state".

Normal choropleth that draws Alaska and Hawaii as insets. 
In addition to a columns named "region" and "value", also requires a column named "state".
}
\section{Super class}{
\code{\link[choroplethr:Choropleth]{choroplethr::Choropleth}} -> \code{USAChoropleth}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-USAChoropleth-render}{\code{USAChoropleth$render()}}
\item \href{#method-USAChoropleth-render_helper}{\code{USAChoropleth$render_helper()}}
\item \href{#method-USAChoropleth-render_state_outline}{\code{USAChoropleth$render_state_outline()}}
\item \href{#method-USAChoropleth-set_zoom}{\code{USAChoropleth$set_zoom()}}
\item \href{#method-USAChoropleth-clone}{\code{USAChoropleth$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="get_ggscale"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-get_ggscale'><code>choroplethr::Choropleth$get_ggscale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="guess_geoid_type"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-guess_geoid_type'><code>choroplethr::Choropleth$guess_geoid_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="choroplethr" data-topic="Choropleth" data-id="initialize"><a href='../../choroplethr/html/Choropleth.html#method-Choropleth-initialize'><code>choroplethr::Choropleth$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-USAChoropleth-render"></a>}}
\if{latex}{\out{\hypertarget{method-USAChoropleth-render}{}}}
\subsection{Method \code{render()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USAChoropleth$render(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-USAChoropleth-render_helper"></a>}}
\if{latex}{\out{\hypertarget{method-USAChoropleth-render_helper}{}}}
\subsection{Method \code{render_helper()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USAChoropleth$render_helper(choropleth.df, scale_name, theme)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-USAChoropleth-render_state_outline"></a>}}
\if{latex}{\out{\hypertarget{method-USAChoropleth-render_state_outline}{}}}
\subsection{Method \code{render_state_outline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USAChoropleth$render_state_outline(states)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-USAChoropleth-set_zoom"></a>}}
\if{latex}{\out{\hypertarget{method-USAChoropleth-set_zoom}{}}}
\subsection{Method \code{set_zoom()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USAChoropleth$set_zoom(zoom)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-USAChoropleth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-USAChoropleth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{USAChoropleth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
