% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{USAChoropleth}
\alias{USAChoropleth}
\title{Normal choropleth that draws Alaska and Hawaii as insets.
In addition to a columns named "region" and "value", also requires a column named "state".}
\format{\preformatted{List of 8
 $ new       :function (...)  
 $ classname : chr "USAChoropleth"
 $ public    :List of 23
  ..$ title               : chr ""
  ..$ legend_name         : chr ""
  ..$ warn                : logi TRUE
  ..$ ggplot_scale        : NULL
  ..$ user.df             : NULL
  ..$ map.df              : NULL
  ..$ choropleth.df       : NULL
  ..$ map.names           : NULL
  ..$ num_buckets         : num 7
  ..$ regions             : NULL
  ..$ discretize          :function ()  
  ..$ bind                :function ()  
  ..$ prepare_map         :function ()  
  ..$ get_scale           :function ()  
  ..$ theme_clean         :function ()  
  ..$ theme_inset         :function ()  
  ..$ format_levels       :function (x, nsep = " to ")  
  ..$ add_state_outline   : logi FALSE
  ..$ initialize          :function (map.df, user.df)  
  ..$ clip                :function ()  
  ..$ render              :function (num_buckets = 7)  
  ..$ render_helper       :function (choropleth.df, scale_name, theme)  
  ..$ render_state_outline:function (states)  
 $ private   : NULL
 $ active    : NULL
 $ inherit   :List of 8
  ..$ new       :function (...)  
  ..$ classname : chr "Choropleth"
  ..$ public    :List of 20
  .. ..$ title        : chr ""
  .. ..$ legend_name  : chr ""
  .. ..$ warn         : logi TRUE
  .. ..$ ggplot_scale : NULL
  .. ..$ initialize   :function (map.df, user.df)  
  .. ..$ render       :function (num_buckets = 7)  
  .. ..$ user.df      : NULL
  .. ..$ map.df       : NULL
  .. ..$ choropleth.df: NULL
  .. ..$ map.names    : NULL
  .. ..$ num_buckets  : num 7
  .. ..$ regions      : NULL
  .. ..$ clip         :function ()  
  .. ..$ discretize   :function ()  
  .. ..$ bind         :function ()  
  .. ..$ prepare_map  :function ()  
  .. ..$ get_scale    :function ()  
  .. ..$ theme_clean  :function ()  
  .. ..$ theme_inset  :function ()  
  .. ..$ format_levels:function (x, nsep = " to ")  
  ..$ private   : NULL
  ..$ active    : NULL
  ..$ inherit   : NULL
  ..$ parent_env:<environment: 0x7fd66ca66e38> 
  ..$ lock      : logi TRUE
  ..- attr(*, "class")= chr "R6ClassGenerator"
 $ parent_env:<environment: 0x7fd66ca66e38> 
 $ lock      : logi TRUE
 - attr(*, "class")= chr "R6ClassGenerator"
}}
\usage{
USAChoropleth
}
\description{
Normal choropleth that draws Alaska and Hawaii as insets.
In addition to a columns named "region" and "value", also requires a column named "state".
}
\keyword{datasets}

