% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa.R
\docType{data}
\name{USAChoropleth}
\alias{USAChoropleth}
\title{Normal choropleth that draws Alaska and Hawaii as insets. 
In addition to a columns named "region" and "value", also requires a column named "state".}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
USAChoropleth
}
\description{
Normal choropleth that draws Alaska and Hawaii as insets. 
In addition to a columns named "region" and "value", also requires a column named "state".
}
\keyword{datasets}
