% Generated by roxygen2 (4.0.1): do not edit by hand
\name{zip_choropleth}
\alias{zip_choropleth}
\title{Create a colored scatterplot, with state borders, of US ZIP codes}
\usage{
zip_choropleth(df, title = "", scaleName = "", num_buckets = 7,
  warn_na = FALSE, states = state.abb, renderAsInsets = TRUE)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".  Region must 5 digit zipcodes as defined in the "zipcode" package.}

\item{title}{An optional title for the map.}

\item{scaleName}{An optional label for the legend.}

\item{num_buckets}{The number of equally sized buckets to places the values in.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many buckets.  For
example, 2 will show values above or below the median, and 9 will show the maximum
resolution.}

\item{warn_na}{If true, choroplethr will emit a warning when a) you give it regions that it is ignoring and b) you do not supply regions that it is rendering.}

\item{states}{A vector of states to render.  Defaults to state.abb.}

\item{renderAsInsets}{If true, Alaska and Hawaii will be rendered as insets on the map.  If false, all 50 states will be rendered
on the same longitude and latitude map to scale. This variable is only checked when the "states" variable is equal to all 50 states.}
}
\value{
A choropleth.
}
\description{
While the function is named zip_choropleth for the same of consistency with other functions in choropleth, the
output is not technically a choropleth map because the borders of the ZIP codes are not drawn.
}
\examples{
library(zipcode)
data(zipcode)
head(zipcode)

df = data.frame(region=zipcode$zip, value=sample.int(nrow(zipcode)))
zip_choropleth(df)
}
\seealso{
The "zipcode" package
}

