% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{zip_map}
\alias{zip_map}
\title{Create a map visualizing US ZIP codes with sensible defaults}
\usage{
zip_map(...)
}
\arguments{
\item{...}{All arguments are ignored.}
}
\description{
This function is deprecated as of choroplethr version 3.0.0. Please use ?zip_choropleth instead. 
The last version of choroplethr 
in which this function worked was version 2.1.1, which can be downloaded from CRAN 
here: http://cran.r-project.org/web/packages/choroplethr/index.html
}
