% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ZipMap}
\alias{ZipMap}
\title{Create a county-level choropleth}
\format{\preformatted{List of 8
 $ new       :function (...)  
 $ classname : chr "CountyChoropleth"
 $ public    :List of 24
  ..$ title               : chr ""
  ..$ legend_name         : chr ""
  ..$ warn                : logi TRUE
  ..$ ggplot_scale        : NULL
  ..$ user.df             : NULL
  ..$ map.df              : NULL
  ..$ choropleth.df       : NULL
  ..$ num_buckets         : num 7
  ..$ set_zoom            :function (zoom)  
  ..$ get_zoom            :function ()  
  ..$ discretize          :function ()  
  ..$ prepare_map         :function ()  
  ..$ theme_clean         :function ()  
  ..$ theme_inset         :function ()  
  ..$ format_levels       :function (x, nsep = " to ")  
  ..$ render_state_outline:function (states)  
  ..$ add_state_outline   : logi TRUE
  ..$ rm_na               : logi TRUE
  ..$ initialize          :function (user.df)  
  ..$ clip                :function ()  
  ..$ bind                :function ()  
  ..$ render              :function (num_buckets = 7)  
  ..$ render_helper       :function (choropleth.df, scale_name, theme)  
  ..$ get_scale           :function ()  
 $ private   :List of 1
  ..$ zoom: NULL
 $ active    : NULL
 $ inherit   :List of 8
  ..$ new       :function (...)  
  ..$ classname : chr "USAChoropleth"
  ..$ public    :List of 23
  .. ..$ title               : chr ""
  .. ..$ legend_name         : chr ""
  .. ..$ warn                : logi TRUE
  .. ..$ ggplot_scale        : NULL
  .. ..$ user.df             : NULL
  .. ..$ map.df              : NULL
  .. ..$ choropleth.df       : NULL
  .. ..$ num_buckets         : num 7
  .. ..$ set_zoom            :function (zoom)  
  .. ..$ get_zoom            :function ()  
  .. ..$ clip                :function ()  
  .. ..$ discretize          :function ()  
  .. ..$ bind                :function ()  
  .. ..$ prepare_map         :function ()  
  .. ..$ get_scale           :function ()  
  .. ..$ theme_clean         :function ()  
  .. ..$ theme_inset         :function ()  
  .. ..$ format_levels       :function (x, nsep = " to ")  
  .. ..$ add_state_outline   : logi FALSE
  .. ..$ initialize          :function (map.df, user.df)  
  .. ..$ render              :function (num_buckets = 7)  
  .. ..$ render_helper       :function (choropleth.df, scale_name, theme)  
  .. ..$ render_state_outline:function (states)  
  ..$ private   :List of 1
  .. ..$ zoom: NULL
  ..$ active    : NULL
  ..$ inherit   :List of 8
  .. ..$ new       :function (...)  
  .. ..$ classname : chr "Choropleth"
  .. ..$ public    :List of 20
  .. .. ..$ title        : chr ""
  .. .. ..$ legend_name  : chr ""
  .. .. ..$ warn         : logi TRUE
  .. .. ..$ ggplot_scale : NULL
  .. .. ..$ initialize   :function (map.df, user.df)  
  .. .. ..$ render       :function (num_buckets = 7)  
  .. .. ..$ user.df      : NULL
  .. .. ..$ map.df       : NULL
  .. .. ..$ choropleth.df: NULL
  .. .. ..$ num_buckets  : num 7
  .. .. ..$ set_zoom     :function (zoom)  
  .. .. ..$ get_zoom     :function ()  
  .. .. ..$ clip         :function ()  
  .. .. ..$ discretize   :function ()  
  .. .. ..$ bind         :function ()  
  .. .. ..$ prepare_map  :function ()  
  .. .. ..$ get_scale    :function ()  
  .. .. ..$ theme_clean  :function ()  
  .. .. ..$ theme_inset  :function ()  
  .. .. ..$ format_levels:function (x, nsep = " to ")  
  .. ..$ private   :List of 1
  .. .. ..$ zoom: NULL
  .. ..$ active    : NULL
  .. ..$ inherit   : NULL
  .. ..$ parent_env:<environment: 0x7fd80896a7c0> 
  .. ..$ lock      : logi TRUE
  .. ..- attr(*, "class")= chr "R6ClassGenerator"
  ..$ parent_env:<environment: 0x7fd80896a7c0> 
  ..$ lock      : logi TRUE
  ..- attr(*, "class")= chr "R6ClassGenerator"
 $ parent_env:<environment: 0x7fd80896a7c0> 
 $ lock      : logi TRUE
 - attr(*, "class")= chr "R6ClassGenerator"
}}
\usage{
ZipMap
}
\description{
Create a county-level choropleth
}
\keyword{datasets}

