% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ZipMap}
\alias{ZipMap}
\title{Create a county-level choropleth}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7f9d9147b6a0> 
 - attr(*, "name")= chr "CountyChoropleth_generator"
}}
\usage{
ZipMap
}
\description{
Create a county-level choropleth
}
\keyword{datasets}

